backend/api/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
httpx==0.25.2

# Redis (for caching and sessions)
redis==5.0.1
hiredis==2.2.3

# AWS SDK (for S3)
boto3==1.29.7

# Image Processing
Pillow==10.1.0
numpy==1.24.4
opencv-python==4.8.1.78
matplotlib==3.8.2

# ML Model Support
torch==2.1.1
torchvision==0.16.1
tensorflow==2.15.0
scikit-learn==1.3.2
onnx==1.15.0
onnxruntime==1.16.3

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic[email]==2.5.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1